f DPLAN 12618 handle layers while generating FaceboundaryWMSUrl:
Currently, every process requires a GIS service called "basemap". This is not always the case in practice. usually A background layer is always set, but this does not have to be the basemap or have the name "basemap".
The new method 'getAvailableGisLayer' will get the available layer to use : available visible and enabled layers are prioritized to be used as available layer, only enabled in the second place then only only visible.

Although a background layer is normally always set, the case where there are no available layers at all is also treated